### 2024-03-21 #6516 @editorialbot
[PRE REVIEW]: Dimensionally Aligned Signal Processing Algorithms Library
**Submitting author:** @jcupsmith (<a href="http://orcid.org/0009-0000-4565-2692">Jacob Smith</a>)
**Repository:** https://github.com/ORNL/dasp-stacker
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.1
