### 2024-03-21 #6516 @danielskatz
👋 @jcupsmith - thanks for this submission. Please add the country to the affiliation in your .md file, and then use the command `@editorialbot generate pdf` to update the pdf. `@editorialbot` commands need to be the first thing in a new comment.

Given that this software is relatively small, and there's basically no commit history, can you say a bit about the effort that went into this software, and if it's based on any previous software.
