+++ paper/paper.md
@@ -24,7 +24,7 @@ authors:
     orcid: 0000-0003-4616-6084
     affiliation: 1
 affiliations:
- - name: Oak Ridge National Laboratory
+ - name: Oak Ridge National Laboratory, USA
    index: 1
 date: 12 October 2023
 bibliography: paper.bib
