### 2024-03-26 #6516 @jcupsmith
@danielskatz below is the commit history from our internal branch for this project, the initial commit being on October 19th, 2022 with 23 additional commits before moving into the repository you see here. This has been an ongoing project for over a year with an unrecorded (but significant) amount of time going into the conceptual understanding of the algorithms, how to properly design visualizations to demonstrate the algorithms, how to effectively implement, internal testing, optimization, presenting to colleagues, peer review, etc. - this to say there is an amount of work that went into this project that is not represented by the commits alone which is hard to properly capture. I am sure you understand, but worth mentioning when attempting to describe the amount of work. If there are other specific metrics you are looking for just let me know!

![image](https://github.com/openjournals/joss-reviews/assets/72476979/b3fb0c80-91ba-47d5-80c2-7c69610002c1)
