### 2024-03-26 #6516 @danielskatz
Thanks @jcupsmith. Given that this isn't clear, and that the amount of code is relatively small, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. You should hear back in a week or two.
