### 2024-05-13 #6516 @kellyrowland
hi @astrogilda @ThomasA 👋 would either of you be interested in and available for reviewing this JOSS submission?
