### 2024-05-28 #6516 @kellyrowland
hi @ixjlyons @vitorsr 👋 would you be interested in and available for reviewing this JOSS submission?
