### 2024-05-29 #6516 @vitorsr
It would be a privilege and a great pleasure to review an ORNL repository. However, before agreeing to review a JOSS submission, I always take a cursory look at the codebase to gauge, among other things, how long it will take to review. This time around there is such a relatively small amount of code that I could almost review it in an hour or so.

While I appreciate the domain-specific terseness and complexities in developing core signal processing methods and algorithms, I feel there could be more in the software engineering side of things. To be clear, I have no doubts about the technical and scientific aspects of the scholarship, and I am certain we can agree research goes much further than lines of code. However as a software artifact there could be a richer library of examples demonstrating properties of the methods, simulation generators and drivers, a more robust and comprehensive test suite, among other things.
