### 2024-05-29 #6516 @kellyrowland
@vitorsr thanks for your look here and the feedback. I will note that the submission has passed a scope review query from JOSS associate editors and so has been deemed appropriate for a publication - it is quite common that development similar to the features you mention takes place through the process of the review, but given the scope review pass I would not want to consider expanded features as a blocker to publication.

Please let me know what your thoughts are. @jcupsmith if you would like to weigh in as well, please do feel free.
