### 2024-06-10 #6516 @kellyrowland
hi @mrava87 @lucaferranti 👋 would you be interested in and available for reviewing this JOSS submission?
