### 2024-06-10 #6866 @lucaferranti
@kellyrowland do you know if the github preview for joss is broken? I've seen it failing fairly consistency recently
