### 2024-06-10 #6866 @kellyrowland
hmm, for viewing the article proof on github?

I'm seeing a rendered PDF at https://github.com/openjournals/joss-papers/blob/joss.06866/joss.06866/10.21105.joss.06866.pdf , but please let me know if you're referring to something else.
