### 2024-06-10 #6866 @lucaferranti
yeah, I'm referring to the article preview, clicking your article I get

![image](https://github.com/openjournals/joss-reviews/assets/49938764/32c5fd68-bfc2-4f30-a257-cbea8f6d929a)
