### 2024-06-10 #6866 @kellyrowland
gotcha, thanks. I haven't seen that as a known or widespread issue, so you might try the usual clearing cookies/cache or even updating Chrome (I'm using Chrome 125.0.6422.142 with the PDF showing up in-browser). less convenient, but the PDF should be viewable via download either way.
