### 2024-07-01 #6866 @kellyrowland
@samiralavi thanks for opening that issue, it looks like it's been resolved. just a ping here on the review status, please let me know if you need to set things down.

likewise @lucaferranti - thanks for getting started on your review, please let me know if you need to set things down.
