### 2024-07-01 ORNL/dasp-stacker#2 @samiralavi
API documentation
There is no documentation that describes the available APIs and modules. The example notebooks are helpful, but a separate API documentation is needed. It could be added to the `README.md` file of the repo.
