### 2024-07-01 ORNL/dasp-stacker#3 @samiralavi
Contribution guide
I couldn't find any contribution guide in the repo, which is required for the JOSS paper to be accepted. The criteria is:

> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

My suggestion is to adding a CONTRIBUTING.md file to repo as it is the common method or adding this content to the `README.md` file of the repo.
