### 2024-07-01 ORNL/dasp-stacker#4 @samiralavi
Example notebook error
I was able to install the Python package both from PyPi and from the repo successfully. However, when running the example notebooks, I found that the `algorithm_walkthrough.ipynb` has issues. The function `sin_creator` is not defined. Looking at the other example, it seems the name has been misspelt and the correct one should be `sine_wave_creator`.
