### 2024-07-01 #6866 @samiralavi
Thanks @kellyrowland for the reminder. I can see that the issue is resolved and I continued with the rest of the review. @jcupsmith  I have created the following issues to be addressed based on the review criteria mentioned above:

https://github.com/ORNL/dasp-stacker/issues/2
https://github.com/ORNL/dasp-stacker/issues/3
https://github.com/ORNL/dasp-stacker/issues/4

I am happy to recommend the paper after these issues are resolved.
