### 2024-08-23 f6f57a1
fixing alg walk and add contributing
