### 2024-08-23 ORNL/dasp-stacker#3 @jcupsmith
added contributing file as requested
