### 2024-08-23 ORNL/dasp-stacker#4 @jcupsmith
@samiralavi I believe I have fixed all of the issues within the algorithm_walkthrough.ipynb file. Just let me know if something is still causing errors for you or if it seems good to go!
