### 2024-09-08 2b2a3cb
edit function docstrings and add to read me (part 1)
