+++ dasp/visualizations.py
@@ -164,6 +164,26 @@ def hasp_update(
     output: widgets.Output,
     alg_type: str = "fixed",
 ):
+    """
+    Updates HASP plots based on the provided input data.
+
+    Args:
+        sample_rate: The sample rate used for the HASP plot.
+        fft: The Fast Fourier Transform (FFT) of the input signal.
+        use_row_norm: Whether to use row normalization for the HASP plot.
+        use_im_norm: Whether to use image normalization for the HASP plot.
+        freq_center: The frequency center used for the HASP plot.
+        bandwidth: The bandwidth used for the HASP plot.
+        max_harmonics: The maximum number of harmonics to display.
+        scale_image: Whether to scale the image for the HASP plot.
+        fig: The figure canvas to use for plotting.
+        ax: The axis to use for plotting.
+        output: The output widget to display the plot in.
+        alg_type: The type of HASP algorithm used (e.g. "Fixed HASP"). Defaults to "fixed".
+
+    Returns:
+        None
+    """
     if alg_type == "fixed":
         hasp_array = hasp_fixed(
             sample_rate=sample_rate,
@@ -216,6 +236,19 @@ def hasp_update(
 
 
 def hasp_viz(signal, freq_center, sample_rate):
+    """
+    Visualizes the HASP algorithm(s).
+
+    Creates an interactive visualization of the HASP algorithm(s) using the provided input data.
+
+    Args:
+        signal: The input signal to visualize.
+        freq_center: The frequency center of the input signal.
+        sample_rate: The sample rate of the input signal.
+
+    Returns:
+        UI: The object containing the visualizations.
+    """
     # create FFT, needed for hasp_fixed call
     fft = gen_fft(signal)
 
@@ -327,6 +360,27 @@ def hasp_viz(signal, freq_center, sample_rate):
 def sine_wave_creator(
     base_freq, sample_rate, duration, noise, wave_type, num_signals, harmonics=0
 ):
+    """
+    Creates a sine wave signal with specified characteristics.
+
+    Generates a sine wave signal with the specified base frequency, sample rate, duration,
+    and noise level. The wave type can be either 'base', 'square', 'sawtooth', or 'triangle', and the signal can be
+    generated with a specified number of harmonics.
+
+    Args:
+        base_freq: The base frequency of the sine wave.
+        sample_rate: The sample rate of the sine wave.
+        duration: The duration of the sine wave.
+        noise: The level of noise to add to the sine wave.
+        wave_type: The type of sine wave to generate ('base', 'square', 'sawtooth', or 'triangle').
+        num_signals: The number of signals to generate.
+        harmonics: The number of harmonics to include in the signal (default=0).
+
+    Returns:
+        out_signal: The generated sine wave signal.
+        time: The associated time array of the generated signal.
+    """
+
     base_freq = base_freq
     sample_rate = sample_rate
     duration = duration
@@ -356,8 +410,9 @@ def sine_wave_creator(
     else:
         out_signal = base_sig
 
+    # recursively add signals based on the num_signals input parameter
     if num_signals > 1:
-        freq_add = int(base_freq + np.sqrt(1))
+        freq_add = int(base_freq + np.sqrt(base_freq))
         other_sig, time = sine_wave_creator(
             freq_add,
             sample_rate,
@@ -369,15 +424,31 @@ def sine_wave_creator(
         )
         out_signal += other_sig
 
-    # pyharmonics
-    # sine of each of the frequencies then add them up
-
     return out_signal, time
 
 
 def sig_update(
     base_freq, sample_rate, duration, harmonics, noise, wave_type, num_signals, fig, ax
 ):
+    """
+    Updates the signal plot with the specified parameters.
+
[…]
