+++ README.md
@@ -2,6 +2,13 @@
 
 [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
 
+## Table of Contents
+1. [Summary](#summary)
+2. [Features/Usage](#features--usage)
+    * [DASP Algorithms](#dasp-algorithms)
+    * [Visualizations](#visualization-functions)
+    * [Other](#other-functions)
+3. [Installation](#installation)
 
 ## Summary
 Dimensionally aligned signal projection (DASP) algorithms are used to analyze fast Fourier transforms (FFTs) and generate visualizations that help focus on the harmonics for specific signals. At a high level, these algorithms extract the FFT segments around each harmonic frequency center, and then align them in equally sized arrays ordered by increasing distance from the base frequency. This allows for a focused view of the harmonic frequencies, which, among other use cases, can enable machine learning algorithms to more easily identify salient patterns. This work seeks to provide an effective open-source implementation of the DASP algorithms as well as functionality to help explore and test how these algorithms work with an interactive dashboard and signal-generation tool.
@@ -12,20 +19,128 @@ Finally, the library provides multiple interactive visualizations, each implemen
 
 A detailed breakdown of how these algorithms work can be seen in 'example/algorithm_walkthrough'
 
-## Features
-* Dashboard-style interactive visualization (Figure 1)
-* Numerical array output for each algorithm
-* Interactive visualization for FASP only
-* Interactive visualization for HASP algorithms (selecting one at a time)
-* Interactive visualization for sine wave generation
+## Features / Usage
 
-## Examples
+Here are some of the main features of the dasp-stacker library:
+* Numerical array output for each algorithm (FASP, Fixed HASP, Decimating HASP, Interpolating HASP).
+* Dashboard-style interactive visualization of HASP output alongside other, more common, signal visualization methods.
+* Custom sine wave generator.
+* Interactive visualization for FASP only.
+* Interactive visualization for HASP algorithms (selecting one at a time).
+* Interactive visualization for sine wave generation.
 
-An example of each feature available, and how to use them, can be seen in 'examples/dasp_example_notebook.ipynb'
+The features of this library can be thought of as 2 seperate sections. One being simple to use functions to provide raw output of the DASP algorithms, which is the main function of this library. As an added set of features, several visualizations are provided to help develop an intuitive understanding of the HASP algorithms, as well as an interactive signal generation function.
+
+The following sections are a breakout of the recommended usage for the available features of the dasp-stacker library. An example of actual function calls, along with the output, can be seen in 'examples/dasp_example_notebook.ipynb'.
+
+### DASP Algorithms
+<hr>
+
+#### FASP
+Generates an STFT.
+
+##### Parameters
+
+* signal (np.ndarray): The raw time domain signal.
+* num_ffts (int): Number of fft slices within the STFT.
+* use_row_norm (bool): Whether to normalize across each FFT individually.
+* use_im_norm (bool): Whether to normalize the entire image.
+
+##### Returns
+* The FASP array -- a 2D NumPy array.
+
+##### Example Usage
+
+    fasp(signal=signal, num_ffts=np.floor(np.sqrt(signal.size / 2)), use_row_norm=True, use_im_norm=True)
+
+#### Fixed HASP
+Fixed Harmonically Aligned Signal Projection generates a HASP array, using a fixed bandwidth around the frequency center.
+
+##### Parameters
+* sample_rate (int): Rate at which the the raw time domain signal was acquired or generated, as samples per second.
[…]
