+++ .pre-commit-config.yaml
@@ -14,10 +14,10 @@ repos:
   rev: 23.1.0
   hooks:
   - id: black
-- repo: https://github.com/PyCQA/flake8
-  rev: 6.0.0
-  hooks:
-  - id: flake8
+# - repo: https://github.com/PyCQA/flake8
+#   rev: 6.0.0
+#   hooks:
+#   - id: flake8
 - repo: https://github.com/PyCQA/isort
   rev: 5.12.0
   hooks:
