### 2024-09-08 d9ee82f
add minor changes to readme
