+++ README.md
@@ -7,7 +7,7 @@
 2. [Features/Usage](#features--usage)
     * [DASP Algorithms](#dasp-algorithms)
     * [Visualizations](#visualization-functions)
-    * [Other](#other-functions)
+    * [Other Functions](#other-functions)
 3. [Installation](#installation)
 
 ## Summary
@@ -70,14 +70,16 @@ Fixed Harmonically Aligned Signal Projection generates a HASP array, using a fix
 
 ##### Example Usage
 
-    hasp_fixed(
-        sample_rate=10_000,
-        fft=fft,
-        bandwidth=1_000,
-        freq_center=1_000,
-        use_row_norm=True,
-        use_im_norm=True,
-        )
+```python
+hasp_fixed(
+    sample_rate=10_000,
+    fft=fft,
+    bandwidth=1_000,
+    freq_center=1_000,
+    use_row_norm=True,
+    use_im_norm=True,
+)
+```
 
 #### Decimating HASP
 Decimating Harmonically Aligned Signal Projection generates a HASP array, allowing bandwidth around the frequency center to grow as the harmonics increase before downsampling each row to the minimum bandwidth size.
@@ -95,16 +97,16 @@ Decimating Harmonically Aligned Signal Projection generates a HASP array, allowi
 * The decimating HASP array -- a 2D NumPy array.
 
 ##### Example Usage
-
-    hasp_decim(
-        sample_rate=10_000,
-        fft=fft,
-        bandwidth=1_000,
-        freq_center=1_000,
-        use_row_norm=True,
-        use_im_norm=True,
-        )
-
+```python
+hasp_decim(
+    sample_rate=10_000,
+    fft=fft,
+    bandwidth=1_000,
+    freq_center=1_000,
+    use_row_norm=True,
+    use_im_norm=True,
+)
+```
 #### Interpolating HASP
 Interpolating Harmonically Aligned Signal Projection generates a HASP array, allowing bandwidth around the frequency center to grow as the harmonics increase before upsampling each row to the maximum bandwidth size.
 
@@ -121,24 +123,147 @@ Interpolating Harmonically Aligned Signal Projection generates a HASP array, all
 * The interpolating HASP array -- a 2D NumPy array.
 
 ##### Example Usage
-
-    hasp_interp(
-        sample_rate=10_000,
-        fft=fft,
-        bandwidth=1_000,
-        freq_center=1_000,
-        use_row_norm=True,
-        use_im_norm=True,
-        )
+```python
+hasp_interp(
+    sample_rate=10_000,
+    fft=fft,
+    bandwidth=1_000,
+    freq_center=1_000,
+    use_row_norm=True,
+    use_im_norm=True,
+    )
+```
 
 ### Visualization Functions
 
---- IN PROGRESS ---
+All visualizations are designed to display inside of a jupyter notebook.
 
+#### HASP Dashboard
+Creates a dashboard of the signal, spectogram, FFT, and HASP output.
+
+This dashboard generates a signal based on the given input and does not accept external signals.
+The purpose of this visualization is to provide common signal visualizations alongside the HASP
+results to demonstrate how a variety of parameters effect the HASP algorithm.
+
+##### Parameters
+
+* base_freq (int): The base frequency of the sine wave, in hertz
+* sample_rate (int): The sample rate of the sine wave as sample per second
+* duration (int): The duration of the sine wave, in seconds.
+* noise (float): The level of noise to add to the sine wave.
+* wave_type (string): The type of sine wave to generate ('base', 'square', 'sawtooth', or 'triangle').
+* bandwidth (int): The bandwidth aroudn the frequency senter to use for the HASP algorithm. To prevent overlap with other harmonics, this value cannot be larger than the base frequency.
[…]
