+++ README.md
@@ -15,7 +15,7 @@ Dimensionally aligned signal projection (DASP) algorithms are used to analyze fa
 
 The DASP library is implemented in Python and contains four types of algorithms for implementing these feature engineering techniques: fixed harmonically aligned signal projection (HASP), decimating HASP, interpolating HASP, and frequency aligned signal projection (FASP). Each algorithm returns a numerical array, which can be visualized as an image. For consistency FASP, which is the terminology used for the short-time Fourier transform, has been implemented as part of the library to provide a similar interface to the STFT of the raw signal. Additionally, the library contains an algorithm to generate artificial signals with basic customizations such as the base frequency, sample rate, duration, number of harmonics, noise, and number of signals.
 
-Finally, the library provides multiple interactive visualizations, each implemented using IPyWidgets and work in a Jupyter environment. A dashboard-style visualization is provided, which contains some common signal-processing visual components (signal, FFT, spectogram) updating in unison with the HASP functions (see Figure 1 below). Seperate from the dashboard, an independent visualization is provided for each of the DASP algorithms as well as the artifical signal generator. These visualizations are included in the library to aid in developing an intuitive understanding how the algorithms are affected by different input signals and parameter selections.
+Finally, the library provides multiple interactive visualizations, each implemented using IPyWidgets and work in a Jupyter environment. A dashboard-style visualization is provided, which contains some common signal-processing visual components (signal, FFT, spectogram) updating in unison with the HASP functions (see screenshot in [Visualizations](#visualization-functions)). Seperate from the dashboard, an independent visualization is provided for each of the DASP algorithms as well as the artifical signal generator. These visualizations are included in the library to aid in developing an intuitive understanding how the algorithms are affected by different input signals and parameter selections.
 
 A detailed breakdown of how these algorithms work can be seen in 'example/algorithm_walkthrough'
 
@@ -36,7 +36,7 @@ The following sections are a breakout of the recommended usage for the available
 ### DASP Algorithms
 <hr>
 
-#### FASP
+#### `fasp()` - Frequency Aligned Signal Projection
 Generates an STFT.
 
 ##### Parameters
@@ -53,7 +53,7 @@ Generates an STFT.
 
     fasp(signal=signal, num_ffts=np.floor(np.sqrt(signal.size / 2)), use_row_norm=True, use_im_norm=True)
 
-#### Fixed HASP
+#### `hasp_fixed()` - Fixed Harmonically Aligned Signal Projection
 Fixed Harmonically Aligned Signal Projection generates a HASP array, using a fixed bandwidth around the frequency center.
 
 ##### Parameters
@@ -81,7 +81,7 @@ hasp_fixed(
 )
 ```
 
-#### Decimating HASP
+#### `hasp_decim()` - Decimating Harmonically Aligned Signal Projection
 Decimating Harmonically Aligned Signal Projection generates a HASP array, allowing bandwidth around the frequency center to grow as the harmonics increase before downsampling each row to the minimum bandwidth size.
 
 ##### Parameters
@@ -107,7 +107,7 @@ hasp_decim(
     use_im_norm=True,
 )
 ```
-#### Interpolating HASP
+#### `hasp_interp()` - Interpolating Harmonically Aligned Signal Projection
 Interpolating Harmonically Aligned Signal Projection generates a HASP array, allowing bandwidth around the frequency center to grow as the harmonics increase before upsampling each row to the maximum bandwidth size.
 
 ##### Parameters
@@ -138,7 +138,7 @@ hasp_interp(
 
[…]
