### 2024-09-09 ORNL/dasp-stacker#2 @jcupsmith
@samiralavi I have added this documentation to the `README.md` file, please take a look to make sure this covers what you are requesting. Thank you for your patience, I wasn't able to get to this until this week.
