### 2024-09-30 #6866 @lucaferranti
Hi @jcupsmith (cc @kellyrowland ) :wave: ,

first of all, my sincere apologies for the long delay in the review. This somehow fell off my notifications list (not a valid excuse of course, my fault).

Overall, I think the work is valuable and should be published. Here are some comments, mainly concerning the repository documentation and its user friendliness.

- **Unit tests**: currently the repository has no github workflow running the unit tests nor instructions on how to run the tests locally. I would strongly recommend to have a github worfklow running the unit tests

- **Installation instructions**: Currently the installation instructions are at the very bottom of the readme, which is very long (see also comment below). While `pip install` is the most obvious way of installing a python package and most people know it, it would be good to move this to the top of the file.

- **Contributing guidelines**: the repository has a `CONTRIBUTING.md` file, which I think it's enough for the minimum requirements. It could be more new contributors friendly to point to those guidelines from the readme

- **Documentation structure**: Currently the composed by the `README` which is pretty long, and the `examples` folder with some jupyter notebooks. **I would recommend to instead leverage sphinx + readthedocs (or similar toolchain) for the documentation.** In the current form, the API documentation is in the README. Not only does this introduce some boilerplate with the docstrings, but it's also proner to go out of sync, while with sphinx one could generate the API documentation from the docstrings. Another advantage of sphinx is that it can build pages from jupyter notebooks, which would allow to integrate the example notebooks into the documentation website. Widgets cannot probably be run interactively of course, but one could have in the generated html page a link to download the notebook or, even better, a binder link to allow the user to run the notebook online. Since generating the html from the notebooks requires to run them, this also simplifies maintainance of the docs and helps preventing the examples in the notebooks go outdated
