### 2024-10-23 #6866 @kellyrowland
Thanks @lucaferranti for the detailed review. You make many good points which I think would ultimately be beneficial for the code repository but are not blockers to the JOSS publication (for example, automated testing is not a requirement so long as there are instructions for testing which are clear to follow and function as expected).
