### 2024-10-24 8ffdcb6
add automated github testing
