### 2024-10-24 32e1fc3
add numpy version of test data
