+++ tests/conftest.py
@@ -1,14 +1,10 @@
 import numpy as np
 import pytest
 
-from examples.h5_handling import get_h5py_data
-
 
 @pytest.fixture(scope="session")
 def flourescent_data():
-    data = get_h5py_data(
-        "./examples/data/fluorescentlights.h5", "/fluorescentlights/1/on", 0, 2_000_000
-    )
+    data = np.load("./tests/data/flourescent_data_1_on.npy")
 
     return data
 
