### 2024-10-24 e2bd3f4
remove 3.10 specific syntax
