### 2024-10-24 #6866 @jcupsmith
@lucaferranti as you suggested, I have added a github workflow for testing. I also added a "Testing" and "Contributing" section to the README, and rearranged so this is all at the top. I will eventually generate documentation using sphinx, but I am leaving that for another day :) Thanks for the feedback!
