+++ README.md
@@ -7,7 +7,8 @@
 1. [Summary](#summary)
 2. [Installation](#installation)
 3. [Testing](#testing)
-4. [Features/Usage](#features--usage)
+4. [Contributing](#contributing)
+5. [Features/Usage](#features--usage)
     * [DASP Algorithms](#dasp-algorithms)
     * [Visualizations](#visualization-functions)
     * [Other Functions](#other-functions)
