### 2024-11-05 #6866 @kellyrowland
@jcupsmith please issue a new tagged release of the software (if changed since the start of the review), and archive it (on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or elsewhere). Then, please then post the version number and archive DOI here in the review issue, and I'll follow some subsequent wrap-up steps from there.
