+++ dasp/__init__.py
@@ -3,4 +3,4 @@
 from dasp import algorithms
 from dasp.algorithms import fasp, hasp_decim, hasp_fixed, hasp_interp
 
-__version__ = "0.1.1"
+__version__ = "0.1.2"
