+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 has context menu
 
+## [0.1.2] - 2024-11-26
+
+* add API documentation
+* fixed issues in example notebook
+
 ## [0.1.1] - 2024-02-12
 
 * fix incorrectly defined dependency in setup.py
