### 2024-11-26 #6866 @jcupsmith
@kellyrowland  I issued a new tagged release (v0.1.2) and archived on Zenodo (DOI -  10.5281/zenodo.14226421)

https://doi.org/10.5281/zenodo.14226421
