### 2024-11-27 #6866 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
