### 2024-11-27 19d5d3d
small changes in paper.md
