### 2024-11-27 #6866 @danielskatz
👋 @jcupsmith - I've suggested a bunch of small changes in https://github.com/ORNL/dasp-stacker/pull/5 - please merge this, or let me know what you disagree with, then we can continue to acceptance and publication.
