### 2024-12-02 #6866 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
