### 2024-12-02 #6866 @danielskatz
Congratulations to  @jcupsmith (Jacob Smith) and co-authors!!

JOSS depends on volunteers and couldn't succeed without you
